# Conflicts:
#	ACore/src/ecflow_version.h
#	ANode/test/TestFlagMigrated.cpp
#	ANode/test/data/migration/default_constructor/Defs.def
#	ANode/test/data/migration/default_constructor/Family.def
#	ANode/test/data/migration/default_constructor/Suite.def
#	ANode/test/data/migration/default_constructor/Task.def
#	ANode/test/data/migration/fixture/boost.checkpt
#	Pyext/src/ExportNodeAttr.cpp

+++ Viewer/src/RepeatEditor.cpp
@@ -122,7 +122,7 @@ void RepeatEditor::buildList(VRepeatAttr *rep)
     int cnt=end-start;
     if(cnt >1)
     {
-        for(size_t i=start; i < end; i++)
+        for(int i=start; i < end; i++)
             modelData_ << QString::fromStdString(rep->value(i));
 
         model_=new QStringListModel(this);
